Building a Micro-service Natively: New files for new section.
These new *.md files and the *.png file support the new
"Buiding Micro-Services Natively" section, which is part of the
"Developer Guides" area in the AGL documentation.  The re-write
will take the place of the existing "Host Configuration"
section.  The re-write was done because the content was mislabeled
as host configuration.  Also, re-written to create a development
flow styled topic.

Change-Id: I4173d9a04fca28bada1aa3d6b97b43e96194f184
Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>