Building Microservice Natively: Added new files
I am adding the six new files for this section, which was
formerly known as "Host Configuration" in the Developer
Guides area of the documentation.  The section is now called
"Building Microservices Natively".  All the files have
been completely rewritten and the PR is from my fork of
agl-documentation and the "sandbox/scottrif29/scottdocs"
branch for consistency.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>